fix(desktop): 保留任务简介中 HTML 图片的替代文字
